meson: Fix gir-to-d call to work in package builds
In 7a99cc5b5a1ae64a3df42d0c22e27e00e0293780, an accidental hardcoded
path for where the gtkd files were was added to the call to gir-to-d.

This broke the Fedora package build, so this change fixes it by
stripping that from the options passed to gir-to-d.